refactoring to fix feedback file issues and better handle submissions consisting of a single source file (rather than a zip of source files)

+++ general/rename.sh
@@ -13,29 +13,31 @@
 ##
 
 shopt -s nullglob
-ALL_FILES="*.zip"
+ALL_FILES="*"
 for old_name in $ALL_FILES; do
-	# chop off "Lab"/"Project"/"Submission" descriptors.
-	new_name=${old_name//[Ll][Aa][Bb]/}
-	new_name=${new_name//[Pp][Rr][Oo][Jj][Ee][Cc][Tt]/}
-	new_name=${new_name//[Pp][Rr][Oo][Jj]/}
-	new_name=${new_name//[Pp][Aa][Rr][Tt]/}
-	new_name=${new_name//[Ss][Uu][Bb][Mm][Ii][Ss][Ss][Ii][Oo][Nn]/}
+	if [ -f "$old_name" ]; then
+		if [ "$old_name" != "grading_guide.txt" ]; then
+			# chop off "Lab"/"Project"/"Submission" descriptors.
+			new_name=${old_name//[Ll][Aa][Bb]/}
+			new_name=${new_name//[Pp][Rr][Oo][Jj][Ee][Cc][Tt]/}
+			new_name=${new_name//[Pp][Rr][Oo][Jj]/}
+			new_name=${new_name//[Pp][Aa][Rr][Tt]/}
+			new_name=${new_name//[Ss][Uu][Bb][Mm][Ii][Ss][Ss][Ii][Oo][Nn]/}
 
-	# Remove the student's username, if present.
-	new_name=${new_name//[a-zA-Z][a-zA-Z][a-zA-Z][0-9][0-9][0-9][0-9]/}
-	new_name=${new_name//[a-zA-Z][a-zA-Z][0-9][0-9][0-9][0-9]/}
-	
-	# Clean the file name for each student submission.
-	# (should be student's name with no #s, -'s, or spaces.)
-	new_name=${new_name//[0-9]/}
-	new_name=${new_name//-/}
-	new_name=${new_name// /}
-	new_name=${new_name//_/}
-	new_name=${new_name//,/_}
+			# Remove the student's username, if present.
+			new_name=${new_name//[a-zA-Z][a-zA-Z][a-zA-Z][0-9][0-9][0-9][0-9]/}
+			new_name=${new_name//[a-zA-Z][a-zA-Z][0-9][0-9][0-9][0-9]/}
+			
+			# Clean the file name for each student submission.
+			# (should be student's name with no #s, -'s, or spaces.)
+			new_name=${new_name//[0-9]/}
+			new_name=${new_name//-/}
+			new_name=${new_name// /}
+			new_name=${new_name//,/_}
 
-	# Copy file and contents into new file and delete old file.
-	cp "$old_name" "$new_name"
-	rm "$old_name"
-
+			# Copy file and contents into new file and delete old file.
+			cp "$old_name" "$new_name"
+			rm "$old_name"
+		fi
+	fi
 done
